Add stripxx.vbs.
This is version N of stripxx.vbs, a script I wrote to remove invalid entries in Sedona library hold lists. The cause of the invalid entries is unclear, but the library hasn't been able to fix them. Hence I wrote this script to suppress them.  It is most fortunate that the Workflows program allows printing to an application, so that I can intercept the holds list before printing it.